chore: restructured logic for generate update and remove
