+++ packages/mockserver-config-writer/src/mockserver-config/index.ts
@@ -17,7 +17,9 @@ export async function generateMockserverConfig(basePath: string, data: Mockserve
     if (!fs) {
         fs = create(createStorage());
     }
-    enhancePackageJson(fs, basePath, data.packageJsonConfig);
+    if (!data.packageJsonConfig?.skip) {
+        enhancePackageJson(fs, basePath, data.packageJsonConfig);
+    }
     await enhanceYaml(fs, basePath, data.webappPath, data.ui5MockYamlConfig);
     return fs;
 }
